+++ src/komponenter/footer/footer-regular/footer-topp/FooterTopp.tsx
@@ -1,6 +1,5 @@
 import React, { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
-import { Heading } from '@navikt/ds-react';
 import Tekst from 'tekster/finn-tekst';
 import { Up } from '@navikt/ds-icons';
 import LenkeMedIkon from 'komponenter/footer/common/lenke-med-ikon/LenkeMedIkon';
@@ -8,39 +7,17 @@ import { AppState } from 'store/reducers';
 import { MenyNode } from 'store/reducers/menu-duck';
 import { findNode, getLanguageNode } from 'utils/meny-storage-utils';
 import BEMHelper from 'utils/bem';
-import { LinksLoader } from '../../../common/content-loaders/LinkLoader';
 import FooterLenker from 'komponenter/footer/common/Lenker';
-import { Locale } from 'store/reducers/language-duck';
 import { DelSkjermLenke } from 'komponenter/footer/common/del-skjerm-lenke/DelSkjermLenke';
+import FooterToppKolonner from './FooterToppKolonner';
 import './FooterTopp.less';
 
 const FooterTopp = () => {
     const cls = BEMHelper('menylinje-topp');
     const { language } = useSelector((state: AppState) => state.language);
-    const context = useSelector((state: AppState) => state.arbeidsflate.status);
     const { data } = useSelector((state: AppState) => state.menypunkt);
-    const [columnsNode, settColumnsNode] = useState<MenyNode>();
     const [personvernNode, settPersonvernNode] = useState<MenyNode>();
     const { PARAMS } = useSelector((state: AppState) => state.environment);
-
-    useEffect(() => {
-        const languageNode = getLanguageNode(language, data);
-        const isLanguageNorwegian = language === Locale.BOKMAL || language === Locale.NYNORSK;
-
-        if (languageNode) {
-            const footerNode = findNode(languageNode, 'Footer');
-            if (footerNode) {
-                const columnsNode = findNode(footerNode, 'Columns');
-                if (columnsNode) {
-                    if (isLanguageNorwegian) {
-                        settColumnsNode(findNode(columnsNode, context));
-                    } else {
-                        settColumnsNode(columnsNode);
-                    }
-                }
-            }
-        }
-    }, [language, context, data, settColumnsNode]);
 
     useEffect(() => {
         const noder = getLanguageNode(language, data);
@@ -57,36 +34,6 @@ const FooterTopp = () => {
             behavior: 'smooth',
         });
         (document.getElementById('top-element') as HTMLElement)?.focus();
-    };
-
-    type FooterToppKolonnerProps = {
-        firstNode: number;
-        numberOfNodes: number;
-    };
-
-    const FooterToppKolonner = (props: FooterToppKolonnerProps) => {
-        const lastNode = props.firstNode + props.numberOfNodes;
-
-        return (
-            <>
-                {columnsNode
-                    ? columnsNode.children.slice(props.firstNode, lastNode).map((columnNode, i) => (
-                          <div key={i} className={'menylenker-seksjon'}>
-                              <Heading level="2" size="small" className="menylenker-overskrift">
-                                  {columnNode.displayName}
-                              </Heading>
-                              <ul>
-                                  <FooterLenker node={columnNode} />
-                              </ul>
-                          </div>
-                      ))
-                    : [...Array(props.numberOfNodes)].map((_, index) => (
-                          <div className={'menylenker-seksjon'} key={index}>
-                              <LinksLoader id={`footer-link-loader-${index}`} />
-                          </div>
-                      ))}
-            </>
-        );
     };
 
     return (
